merge with master (#1)
* Create control_climate.py

A simple script for controlling the climate. Can be used with input arguments "start" or "stop" and needs no paths to config files etc. It is easily called by either Node-red or Domoticz etc.

* Adding SOC (state_of_charge) to BatteryStatus

Note: apparently only Leaf 2016 has SOC in it's data

* Adding SOC/state_of_charge to example

* Started on location info

* Addedd location

* Adding OpenStreetMap url to example output

Ctrl-click in terminal to open a browser to show you the map